reorder a bit the 'draw' signal hack
we draw the label in the box 'draw' signal, but we return FALSE, so that if there are other elements in the box they do get drawn too.
we also actually pass the label to the callback instead of looking for it.
this allows us to put a widget a little more complex in the same spot.

Signed-off-by: Niv Sardi <xaiki@debian.org>